+++ src/charade/components/charade-playground/charade-playground.tsx
@@ -1,10 +1,13 @@
 import * as React from 'react'
 import * as R from 'ramda'
+import { Modalize } from 'react-native-modalize'
 import { CustomKeyboard } from '@core/custom-keyboard/custom-keyboard'
 import { CLEAR_BUTTON_ID, SEND_BUTTON_ID } from '@core/custom-keyboard/custom-keyboard.constants'
 import { useForceUpdate } from '@core/hooks/use-force-update.hook'
 import { CharadePlaygroundRow, CharadePlaygroundRowsList } from './charade-playground.styled'
 import { CharadeField } from '../charade-field/charade-field'
+import { createNewContents, updateRGYLetters } from '../../helpers'
+import { EndModal } from '../end-modal/end-modal'
 
 interface Props {
   word: string;
@@ -13,7 +16,9 @@ interface Props {
 
 export const CharadePlayground = ({ word, allWords }: Props) => {
   const count = word.length
+  const modalizeRef = React.useRef<Modalize | null>(null)
 
+  const [ success, setSuccess ] = React.useState<boolean>(false)
   const [ isError, setError ] = React.useState<boolean>(false)
   const [ activeRow, setActiveRow ] = React.useState(0)
   const [ activeIndex, setActiveIndex ] = React.useState(0)
@@ -36,60 +41,40 @@ export const CharadePlayground = ({ word, allWords }: Props) => {
         return
       }
 
-      const letters = R.map(
-        (index: number) => contents[index + (activeRow * (count + 1))]
-      )(R.times(R.identity, count)) as string[]
-
-      let newGreenLetters: string[] = [ ...greenLetters ]
-      let newYellowLetters: string[] = [ ...yellowLetters ]
-      let newRedLetters: string[] = [ ...redLetters ]
-
-      letters.forEach((l: string, i: number) => {
-        if (word.includes(l)) {
-          if (word[i] === l) {
-            newGreenLetters = [ ...newGreenLetters, l ]
-          } else {
-            newYellowLetters = [ ...newYellowLetters, l ]
-          }
-        } else {
-          newRedLetters = [ ...newRedLetters, l ]
-        }
-      })
+      const [ newRedLetters, newGreenLetters, newYellowLetters ] = updateRGYLetters(
+        contents, activeRow, count, word, [ redLetters, greenLetters, yellowLetters ],
+      )
 
       setGreenLetters(newGreenLetters)
       setYellowLetters(newYellowLetters)
       setRedLetters(newRedLetters)
-
       setActiveRow(R.inc)
       setActiveIndex(0)
-    } else if (letter === CLEAR_BUTTON_ID) {
-      setContents((oldContents: (string | undefined)[]) => {
-        let newContents = oldContents
-        newContents[activeIndex + (activeRow * (count + 1))] = undefined
-        return newContents
-      })
 
-      if (activeIndex > 0) {
-        setActiveIndex(activeIndex - 1)
+      if ([ ...new Set(newGreenLetters) ].length === word.length) {
+        setSuccess(true)
+        setTimeout(() => {
+          modalizeRef?.current?.open()
+        }, 1000)
       }
 
+      if (activeRow === 5) {
+        setTimeout(() => {
+          modalizeRef?.current?.open()
+        }, 1000)
+      }
+    } else if (letter === CLEAR_BUTTON_ID) {
+      setContents(createNewContents(activeIndex, activeRow, count, undefined))
+      if (activeIndex > 0) setActiveIndex(activeIndex - 1)
       forceUpdate()
     } else {
-      setContents((oldContents: (string | undefined)[]) => {
-        let newContents = oldContents
-        newContents[activeIndex + (activeRow * (count + 1))] = letter
-        return newContents
-      })
-
-      if (activeIndex + 1 < count) {
-        setActiveIndex(activeIndex + 1)
-      }
-
+      setContents(createNewContents(activeIndex, activeRow, count, letter))
+      if (activeIndex + 1 < count) setActiveIndex(activeIndex + 1)
       forceUpdate()
     }
 
     setError(false)
-  }, [ contents, activeIndex, activeRow, isError ])
+  }, [ contents, activeIndex, activeRow, isError, greenLetters ])
 
   const handlePress = (index: number, rowIndex: number) => () => {
     if (rowIndex === activeRow) {
@@ -121,6 +106,8 @@ export const CharadePlayground = ({ word, allWords }: Props) => {
         onPress={onPressLetter}
         {...{ greenLetters, yellowLetters, redLetters }}
       />
+
+      <EndModal {...{ success, word, modalizeRef }} />
     </>
   )
 }