+++ src/charade/helpers/index.ts
@@ -1,0 +1,2 @@
+export { createNewContents } from './create-new-contents.helper'
+export { updateRGYLetters } from './update-rgy-letters.helper'