+++ src/developer/developer.styled.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components/native'
+
+export const DeveloperContainer = styled.View`
+  padding-horizontal: 10px;
+`