+++ src/more/hooks/use-more-options.hook.ts
@@ -5,12 +5,8 @@ import { Modalize } from 'react-native-modalize'
 import { useTheme } from 'styled-components/native'
 import { useNavigation } from '@react-navigation/native'
 import { O } from '_otils'
-import { NumberFlag } from '@core/models'
 import { ThemeModel } from '@core/styled/models'
 import { COLOR } from '@core/colors/colors.constants'
-import { clearSearchHistoryAlert } from '@core/alerts/clear-seearch-history-alert'
-import { Storage } from '@core/storage/storage'
-import { STORAGE_KEY } from '@core/storage/storage.constants'
 import { deactivatePremiumAlert } from '@core/alerts/deactivate-premium-alert'
 import { premiumService } from '@core/premium-service/premium-service'
 import { useLocalize } from '@core/hooks/use-localize.hook'
@@ -20,15 +16,13 @@ import { MoreOption } from '../more.models'
 import { darkThemeEnabledSelector, hapticFeedbackEnabledSelector, languageCodeSelector } from '../../settings/store/settings.selectors'
 import { SCREEN } from '../../navigation/navigation.constants'
 import {
-  setDarkThemeEnabledAction,
-  setHapticFeedbackEnabledAction, setLanguageCodeAction, setNativeSearchEngineEnabledAction,
+  setDarkThemeEnabledAction, setHapticFeedbackEnabledAction, setLanguageCodeAction,
 } from '../../settings/store/settings.slice'
 
 type Options = [
   MoreOption<undefined>,
   MoreOption<LANGUAGE_CODES>,
   MoreOption<0 | 1 | -1>,
-  MoreOption<boolean>,
   MoreOption<boolean>,
   MoreOption<undefined>,
   MoreOption<undefined>,
@@ -44,7 +38,6 @@ interface UseMoreOptions {
 
 export const useMoreOptions = (
   premiumModalRef: React.MutableRefObject<Modalize | null>,
-  nativeSearchEngineEnabled: NumberFlag,
   premium: number,
 ): UseMoreOptions => {
   const theme = useTheme() as ThemeModel
@@ -58,22 +51,13 @@ export const useMoreOptions = (
 
   const isPending = R.any(
     R.isNil,
-    [ hapticFeedbackEnabled, nativeSearchEngineEnabled, premium ],
+    [ hapticFeedbackEnabled, premium ],
   )
 
   const isPremium = premium > 0
 
   const handleChangeHapticFeedback = (_hapticFeedbackEnabled: boolean) =>
     dispatch(setHapticFeedbackEnabledAction(O.toNumberFlag(_hapticFeedbackEnabled)))
-
-  const handleChangeNativeSearchEngine = (_nativeSearchEngineEnabled: boolean) =>
-    dispatch(setNativeSearchEngineEnabledAction(O.toNumberFlag(_nativeSearchEngineEnabled)))
-
-  const handleClearSearchHistory = () => {
-    clearSearchHistoryAlert(() => {
-      Storage.set(STORAGE_KEY.SEARCH_RESULT, JSON.stringify([]))
-    })
-  }
 
   const handleChangeTheme = (value: 0 | 1 | -1) =>
     dispatch(setDarkThemeEnabledAction(value))
@@ -120,16 +104,6 @@ export const useMoreOptions = (
         onChange: handleChangeHapticFeedback,
       },
       {
-        title: localize().native_search_engine,
-        value: !!nativeSearchEngineEnabled,
-        onChange: handleChangeNativeSearchEngine,
-      },
-      {
-        title: localize().clear_search_history,
-        onChange: handleClearSearchHistory,
-        icon: 'delete',
-      },
-      {
         title: 'Scrabble Mania',
         onChange: () => navigation.navigate(SCREEN.MORE_MANIA),
         icon: 'web',
@@ -145,11 +119,17 @@ export const useMoreOptions = (
         icon: 'help',
       },
       {
+        title: localize().advanced_settings,
+        onChange: () => navigation.navigate(SCREEN.DEVELOPER),
+        icon: 'wrench',
+        // TODO: hidden: !__DEV__,
+      },
+      {
         title: localize().about_author,
         onChange: () => navigation.navigate(SCREEN.MORE_AUTHOR),
         icon: 'account-question',
-      }
-    ], [ hapticFeedbackEnabled, nativeSearchEngineEnabled, premium, isPending, languageCode, darkThemeEnabled ])
+      },
+    ], [ hapticFeedbackEnabled, premium, isPending, languageCode, darkThemeEnabled ])
 
     return { handleDeactivatePremium, getOptions }
 }